FaceRecognitionAccessControl-main/requirements.txt
numpy
opencv-python
tk
pillow
face-recognition
cmake
dlib